hmdfs: modify read and write interface in hmdfs ohos inclusion category: bugfix issue: #I4T9TP CVE: NA
----------------------------------------------

vfs_read and vfs_write is not exported, change interface to
vfs_read_iter and vfs_write_iter

+++ fs/hmdfs/comm/connection.h
@@ -144,7 +144,7 @@ struct hmdfs_peer {
 	uint64_t device_id;
 	unsigned long conn_time;
 	uint8_t version;
-	u8 status;
+	int status;
 	u64 features;
 	long long old_sb_dirty_count;
 	atomic64_t sb_dirty_count;
